Move field tests to field package
