+++ src/test/java/seedu/address/model/field/EmailTest.java
@@ -1,12 +1,10 @@
-package seedu.address.model.property;
+package seedu.address.model.field;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
-
-import seedu.address.model.field.Email;
 
 public class EmailTest {
